requirements.txt
boto3==1.17.22
matplotlib==3.3.4
numpy==1.20.1
h5py==3.2.1
Pillow==8.1.0
pytorch-lightning==1.1.4
wandb==0.10.21
PyYAML==5.3.1
Shapely==1.7.1
numba==0.52.0
test-tube==0.7.3
torch==1.7.1+cu110
tqdm==4.59.0
s3fs==2021.4.0